Move Markdown-specific doctest code into submodule
